requirements.txt
Django~=3.2.8
pip~=21.2.3
pytz~=2021.3
sqlparse~=0.4.2
asgiref~=3.4.1
setuptools~=57.4.0
djangorestframework~=3.12.4
mysqlclient~=2.0.3